k-food-helpdesk/indexer/requirements.txt
openai
psycopg2-binary
python-dotenv
pandas